add tsconfig and update jsdoc comment
